# Introduce the batch scan to improve the performance of the query and limit the memory usage