### Significant Decrease in Memory Usage

From **940.42MB** to **195.27MB** total

### Improved Latency

```markdown

| **Latency**           | N   | Min   | Median | Mean | StdDev | Max    |
|-----------------------|-----|-------|--------|------|--------|--------|
| **Old**               | 100 | 2.5ms | 3.2ms  | 4ms  | 2.4ms  | 16.9ms |
| **New**               | 100 | 1.6ms | 2ms    | 2.2ms| 600µs  | 4.3ms  |
```


- [x] If this pull request closes/resolves/fixes an existing issue, replace the issue number. Fixes apache/skywalking#12913.
- [x] Update the [`CHANGES` log](https://github.com/apache/skywalking-banyandb/blob/main/CHANGES.md).